backend: 80% ported to arangodb, redis portable.
Also began "abstracting" the db library, so any changes
to it in the future do not require the intense amount
of work this has been, and instead only require a few
changes to the DB.

Redis has been fully made "portable", AFAIK that just
means it won't fail outside of a docker container now.

Fixed a lot of issues with the DB by rewriting so far,
and made unknown input or 0 results to be an error.